Move TaskHeaderInfo to admin's components
