+++ src/pages/Admin/Home/Dashboard/TaskDetail/Applicant/ApplicantDetailInfo/UserDetails/component.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from "react";
 import PersonOutlinedIcon from "@material-ui/icons/PersonOutlined";
 import EmailOutlinedIcon from "@material-ui/icons/EmailOutlined";
-import { TaskHeaderInfo } from "../../../TaskHeaderInfo";
+import { TaskHeaderInfo } from "../../../../../../components/TaskHeaderInfo";
 import { IUserDetailsProps } from "./interfaces";
 import styles from "./styles.module.scss";
